backend/requirements.txt
fastapi==0.115.6              # FastAPI 프레임워크
uvicorn[standard]==0.34.0     # ASGI 서버 (FastAPI 실행용)
python-dotenv==1.0.1          # .env 파일을 통한 환경 변수 관리
cx_Oracle==8.3.0              # Oracle 데이터베이스용 Python 클라이언트 라이브러리
sqlalchemy==2.0.37            # ORM
huggingface_hub==0.27.1       # HuggingFace 라이브러리
passlib==1.7.4
PyJWT==2.10.1